requirements.txt
numpy>=1.26.1
opencv-python>=4.8.1
matplotlib>=3.8.1
scipy>=1.11.3
tqdm>=4.66.1
pyyaml